Implement copying of sql data into internal memory model.
There now exists a reasonable architectural boundary between the database
backend (currently limited to Sqlite3) and the internal data structures.
This should provide some flexibility going forward, particularly wrt
integrating other backends.

Data are no longer displayed on screen in this commit.  In theory, it
should be straight forward to render the models, but the new data
modeling is untested.  Crashes are expected.