Write yml converter for cyaml.
Because cyaml doesn't handle dynamic keys (so far as I know), and
because my format for representing a record is quite simple, I've
written a processor that converts the record yml (as it will be seen by
the end user) to a format that the cyaml parser can handle.

Cyaml's validation can take over from there.

This would probably be better handled by cyaml itself, so that error
messages can be interpreted reasonably, but this will do for now.