Implement loading of multiple tables and fk relationships.
Tables are now loaded into a vector as they are needed.  There is
currently no way to see what tables are loaded in the UI.

If a column has a foreign key relationship, the table related to that
column is loaded so that its related record can be linked to the
dependent table.

Code has also been reorganised.