Add additional metadata to loaded tables.
I now gather additional column info, especially primary key status,
and nullability, which is required to establish whether or not the
primary key should be read-only.

I've also moved the column type into the Data_Column struct and no
longer check the type for every iteration of the column data.  This
was a redundant step and wasted CPU cycles.

I still copy the data type to the cell for the time being, but it may be
just as easy to reference this info from the containing column.